chmod -x
I double-clicked on one of these expecting to open it in a text editor, but instead it created a bunch of files in the current directory. The "x" access permission makes Linux treat it as a bash script, and I guess it went through the Haxe file and executed the lines that were valid bash.